+++ src/main/java/net/cyanmarine/simple_veinminer/SimpleVeinminer.java
@@ -6,6 +6,10 @@ import net.cyanmarine.simple_veinminer.server.SimpleVeinminerServer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
+import net.fabricmc.fabric.api.gamerule.v1.GameRuleFactory;
+import net.fabricmc.fabric.api.gamerule.v1.GameRuleRegistry;
+import net.fabricmc.fabric.api.message.v1.ServerMessageEvents;
+import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
 import net.fabricmc.fabric.api.networking.v1.PlayerLookup;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
@@ -13,10 +17,7 @@ import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ShearsItem;
-import net.minecraft.item.SwordItem;
-import net.minecraft.item.ToolItem;
+import net.minecraft.item.*;
 import net.minecraft.network.PacketByteBuf;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -26,6 +27,7 @@ import net.minecraft.text.TranslatableTextContent;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.registry.Registry;
+import net.minecraft.world.GameRules;
 import net.minecraft.world.World;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -35,9 +37,24 @@ import java.util.Collection;
 import java.util.UUID;
 
 public class SimpleVeinminer implements ModInitializer {
+    public static final String MOD_ID = "simple_veinminer";
     public static final Logger LOGGER = LoggerFactory.getLogger("SimpleVeinMiner");
     private static ArrayList<UUID> playersVeinMining;
     static MinecraftServer server;
+    public static final GameRules.Key<GameRules.BooleanRule> SERVER_SIDE_VEINMINING = GameRuleRegistry.register("doServerSideVeinmining", GameRules.Category.MISC, GameRuleFactory.createBooleanRule(false , (server, rule) -> {
+        Collection<ServerPlayerEntity> players = PlayerLookup.all(server);
+
+        boolean newValue = rule.get();
+
+        for (ServerPlayerEntity player : players) {
+            if (newValue)
+                player.sendMessage(Text.of("Server side veinmining was turned on. Mining while shifting will veinmine"));
+            else
+                player.sendMessage(Text.of("Server side veinmining was turned off"));
+
+            sendServerSideVeinminingUpdate(player, newValue);
+        }
+    }));
 
     public static ArrayList<BlockPos> getBlocksToVeinmine(World world, BlockPos pos, BlockState state, int maxBlocks) {
         Block compareTo = state.getBlock();
@@ -58,6 +75,24 @@ public class SimpleVeinminer implements ModInitializer {
         }
 
         return blocksToBreak;
+    }
+
+    public static int getMaxBlocks(Item item) {
+        SimpleConfig config = SimpleVeinminer.getConfig();
+
+        int maxBlocks = config.limits.maxBlocks;
+        if (config.limits.materialBasedLimits) {
+            int limitingFactor = 6;
+
+            if (item instanceof ToolItem)
+                limitingFactor -= ((ToolItem)item).getMaterial().getMiningLevel() + 1;
+            else if (item instanceof ShearsItem)
+                limitingFactor -= ToolMaterials.IRON.getMiningLevel() + 1;
+
+            maxBlocks = maxBlocks / Math.max(1, limitingFactor);
+        }
+
+        return  maxBlocks;
     }
 
     private static boolean listIncludes(Block block, SimpleConfig.Restrictions restrictions) {
@@ -115,11 +150,27 @@ public class SimpleVeinminer implements ModInitializer {
             server.execute(()->setVeinmining(player, isVeinMining));
         });
 
-        ServerPlayConnectionEvents.JOIN.register((handler, sender, server)->syncConfig(handler.player));
+        ServerPlayConnectionEvents.JOIN.register((handler, sender, server)-> {
+            boolean serverSideVeinmining = server.getGameRules().getBoolean(SERVER_SIDE_VEINMINING);
+            ServerPlayerEntity player = handler.getPlayer();
+
+            sendServerSideVeinminingUpdate(player, serverSideVeinmining);
+
+            if (serverSideVeinmining)
+                player.sendMessage(Text.translatable("This server has server side veinmining turned on. Mining while shifting will veinmine"));
+
+            syncConfig(player);
+        });
         ServerPlayConnectionEvents.DISCONNECT.register((handler, server1)->setVeinmining(handler.getPlayer(), false));
         ServerLifecycleEvents.SERVER_STARTED.register(server -> { SimpleVeinminer.server = server; LOGGER.info("New server");});
 
         LOGGER.info("Simple VeinMiner initialized");
+    }
+
+    public static void sendServerSideVeinminingUpdate(ServerPlayerEntity player, boolean value) {
+        PacketByteBuf buf = PacketByteBufs.create();
+        buf.writeBoolean(value);
+        ServerPlayNetworking.send(player, Constants.SERVERSIDE_UPDATE, buf);
     }
 
     public static void syncConfig(ServerPlayerEntity player) {
@@ -134,7 +185,7 @@ public class SimpleVeinminer implements ModInitializer {
     }
 
     public static boolean isVeinmining(PlayerEntity player) {
-        return playersVeinMining.contains(player.getUuid());
+        return (player.isSneaking() && player.world.getGameRules().getBoolean(SERVER_SIDE_VEINMINING)) || playersVeinMining.contains(player.getUuid());
     }
 
     private static void setVeinmining(PlayerEntity player, boolean isVeinMining) {